% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDown.R
\name{modelDown}
\alias{modelDown}
\title{Generates a website with HTML summaries for predictive models}
\usage{
modelDown(..., modules = c("auditor", "drifter", "model_performance",
  "variable_importance", "variable_response"), output_folder = "output",
  repository_name = "repository", should_open_website = TRUE)
}
\arguments{
\item{...}{one or more explainers created with \code{DALEX::explain()} function. Pair of explainer could be provided to check drift of models}

\item{modules}{modules that should be included in the website}

\item{output_folder}{folder where the website will be saved}

\item{repository_name}{name of local archivist repository that will be created}

\item{should_open_website}{should generated website be automatically opened in default browser}
}
\description{
Generates a website with HTML summaries for predictive models
}
\details{
Additional arguments that could by passed by name:
\itemize{
  \item{remote_repository_path} {Path to remote repository that stores folder with archivist repository. If not provided, links to local repository will be shown.}
  \item{device} {Device to use. Tested for "png" and "svg", but values from \code{ggplot2::ggsave} function should be working fine. Defaults to "png".}
  \item{vr.vars} {variables which will be examined in Variable Response module. Defaults to all variables. Example vr.vars = c("var1", "var2")}
  \item{vr.type} {types of examinations which will be conducteed in Variable Response module. Defaults to "pdp". Example vr.type = c("ale", "pdp")}
}
}
\author{
Przemysław Biecek, Magda Tatarynowicz, Kamil Romaszko, Mateusz Urbański
}
