% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDown.R
\name{modelDown}
\alias{modelDown}
\title{Create page with summary for your models}
\usage{
modelDown(..., modules = c("model_performance", "variable_importance",
  "variable_response", "prediction_breakdown"), output_folder = "output")
}
\arguments{
\item{...}{explainers}

\item{modules}{selected modules to generate}

\item{output_folder}{folder where generated page will be stored}
}
\description{
Create page with summary for your models
}
\examples{
\dontrun{
require(ranger)
require(breakDown)
# ranger
HR_ranger_model <- ranger(as.factor(left) ~ .,
                      data = HR_data, num.trees = 500, classification = TRUE, probability = TRUE)
explainer_ranger <- explain(HR_ranger_model,
                      data = HR_data, y = HR_data$left, function(model, data) {
 return(predict(model, data)$prediction[,2])
}, na.rm=TRUE)

# glm
HR_glm_model <- glm(left~., HR_data, family = "binomial")
explainer_glm <- explain(HR_glm_model, data=HR_data, y = HR_data$left)

modelDown::modelDown(explainer_ranger, explainer_glm) #all defaults
}
}
\author{
Magda Tatarynowicz, Kamil Romaszko, Mateusz Urabński
}
