% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDown.R
\name{modelDown}
\alias{modelDown}
\title{Create page with summary for your models}
\usage{
modelDown(..., modules = c("model_performance", "variable_importance",
  "variable_response", "prediction_breakdown"), output_folder = "output")
}
\arguments{
\item{...}{explainers}

\item{modules}{selected modules to generate}

\item{output_folder}{folder where generated page will be stored}
}
\description{
Create page with summary for your models
}
\details{
Additional argument that could by passed by name:
\itemize{
  \item{vr.vars} {variables which will be examined in Variable Response module. Defaults to all variables. Example vr.vars = c("var1", "var2")}
  \item{pb.observations} {observations which will be examined in Prediction Breakdown module. When not given it selects worst predicted observations for each model. Example pb.observations = c(1,2,3) where 1,2,3 are observation numbers.}
  \item{vr.type} {types of examinations which will be conducteed in Variable Response module. Defaults to c("ale", "pdb"). Example vr.type = "ale"}
  \item{plot_width default} {width for plots. Defaults to 800. Example plot_width = 750}
  \item{vr.plot_width} {Override plot width for Variable Response module. Defaults to plot_width. Example vr.plot_width = 750}
  \item{mp.plot_width} {Override plot width for Model Performance module. Defaults to plot_width. Example mp.plot_width = 750}
  \item{pb.plot_width} {Override plot width for Prediction Breakdown module. Defaults to plot_width. Example pb.plot_width = 750}
  \item{vi.plot_width} {Override plot width for Variable Importance module. Defaults to plot_width. Example vi.plot_width = 750}
  \item{font_size default} {width for plots. Defaults is calculated basing on plot width. Plot with width 800 has font size of 19. Example font_size = 25}
  \item{vr.font_size} {Override font size for Variable Response module. Defaults to font_size. Example vr.font_size = 25}
  \item{mp.font_size} {Override font size for Model Performance module. Defaults to font_size. Example mp.font_size = 25}
  \item{pb.font_size} {Override font size for Prediction Breakdown module. Defaults to font_size. Example pb.font_size = 25}
  \item{vi.font_size} {Override font size for Variable Importance module. Defaults to font_size. Example vi.font_size = 25}
}
}
\examples{
\dontrun{
require(ranger)
require(breakDown)
require(DALEX)

# ranger
HR_ranger_model <- ranger(as.factor(left) ~ .,
                      data = HR_data, num.trees = 500, classification = TRUE, probability = TRUE)
explainer_ranger <- explain(HR_ranger_model,
                      data = HR_data, y = HR_data$left, function(model, data) {
 return(predict(model, data)$prediction[,2])
}, na.rm=TRUE)

# glm
HR_glm_model <- glm(left~., HR_data, family = "binomial")
explainer_glm <- explain(HR_glm_model, data=HR_data, y = HR_data$left)

modelDown::modelDown(explainer_ranger, explainer_glm) #all defaults

modelDown::modelDown(explainer_glm,
  modules = c("model_performance", "variable_importance",
              "variable_response", "prediction_breakdown"),
  output_folder = "modelDown_output",
  vr.vars= c("average_montly_hours", "time_spend_company"),
  pb.observations = c(1,2,3),
  vr.type = "ale",
  plot_width=700,
  pb.plot_width=810,
  mp.plot_width=820,
  vi.plot_width=830,
  vr.plot_width=840
  font_size=15,
  pb.font_size=20,
  mp.font_size=20,
  vi.font_size=20,
  vr.font_size=20)
}
}
\author{
Magda Tatarynowicz, Kamil Romaszko, Mateusz Urabński
}
