% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDown.R
\name{modelDown}
\alias{modelDown}
\title{Generates a website with HTML summaries for predictive models}
\usage{
modelDown(..., modules = c("model_performance", "variable_importance",
  "variable_response", "prediction_breakdown"), output_folder = "output")
}
\arguments{
\item{...}{one or more explainers createdwith \code{DALEX::explain()} function}

\item{modules}{modules that should be included in the website}

\item{output_folder}{folder where the website will be saved}
}
\description{
Generates a website with HTML summaries for predictive models
}
\details{
Additional arguments that could by passed by name:
\itemize{
  \item{vr.vars} {variables which will be examined in Variable Response module. Defaults to all variables. Example vr.vars = c("var1", "var2")}
  \item{pb.observations} {observations which will be examined in Prediction Breakdown module. When not given it selects worst predicted observations for each model. Example pb.observations = c(1,2,3) where 1,2,3 are observation numbers.}
  \item{vr.type} {types of examinations which will be conducteed in Variable Response module. Defaults to "pdp". Example vr.type = c("ale", "pdp")}
  \item{plot_width default} {width for plots (in inches). Defaults to 8. Example plot_width = 750}
  \item{vr.plot_width} {Override plot width for Variable Response module. Defaults to plot_width. Example vr.plot_width = 750}
  \item{mp.plot_width} {Override plot width for Model Performance module. Defaults to plot_width. Example mp.plot_width = 750}
  \item{pb.plot_width} {Override plot width for Prediction Breakdown module. Defaults to plot_width. Example pb.plot_width = 750}
  \item{vi.plot_width} {Override plot width for Variable Importance module. Defaults to plot_width. Example vi.plot_width = 750}
}
}
\examples{
\dontrun{
require("ranger")
require("breakDown")
require("DALEX")

# ranger
HR_ranger_model <- ranger(as.factor(left) ~ .,
                      data = HR_data, num.trees = 500, classification = TRUE, probability = TRUE)
explainer_ranger <- explain(HR_ranger_model,
                      data = HR_data, y = HR_data$left, function(model, data) {
 return(predict(model, data)$prediction[,2])
}, na.rm=TRUE)

# glm
HR_glm_model <- glm(left~., HR_data, family = "binomial")
explainer_glm <- explain(HR_glm_model, data=HR_data, y = HR_data$left)

modelDown::modelDown(explainer_ranger, explainer_glm) #all defaults

modelDown::modelDown(explainer_glm,
  modules = c("model_performance", "variable_importance",
              "variable_response", "prediction_breakdown"),
  output_folder = "modelDown_output",
  vr.vars= c("average_montly_hours", "time_spend_company"),
  pb.observations = c(1,2,3),
  vr.type = "ale",
  plot_width=700,
  pb.plot_width=810,
  mp.plot_width=820,
  vi.plot_width=830,
  vr.plot_width=840)
}
}
\author{
Magda Tatarynowicz, Kamil Romaszko, Mateusz Urabński
}
